# Precompute encodeWithKeys buffer size to avoid resizes