@howardjohn 2021-09-14
I don't think the test failure is from my test. Maybe its a flaky test? Its in the tracing code, not stats.